% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_primers.R
\name{count_primers}
\alias{count_primers}
\title{Count primers}
\usage{
count_primers(r1_files, fwd_primer, rev_primer, r2_files = NULL)
}
\arguments{
\item{r1_files}{list of fastq files with the forward read}

\item{fwd_primer}{forward primer sequence as a string}

\item{rev_primer}{reverse primer sequence as a string}

\item{r2_files}{list of fastq files with the reverse read, optional}
}
\value{
a tibble with the counts of the primers hits in different orientations, one row per file
}
\description{
This function counts the number of times it finds a match to the given primer in a fastq file
}
