% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getN}
\alias{getN}
\title{Get number of uniques sequences from a dada2 object}
\usage{
getN(x)
}
\arguments{
\item{x}{object to get unique sequences from.  See \code{\link[dada2]{getUniques}} for details.}
}
\description{
Get number of uniques sequences from a dada2 object
}
